refactor(lib): extract `enhanceConfiguration` into utils
