+++ src/lib/utils.js
@@ -8,6 +8,9 @@ import uniq from 'lodash/uniq';
 import mapKeys from 'lodash/mapKeys';
 import mapValues from 'lodash/mapValues';
 import curry from 'lodash/curry';
+import mergeWith from 'lodash/mergeWith';
+import union from 'lodash/union';
+import isPlainObject from 'lodash/isPlainObject';
 import hogan from 'hogan.js';
 
 export {
@@ -27,6 +30,7 @@ export {
   isReactElement,
   deprecate,
   parseAroundLatLngFromString,
+  enhanceConfiguration,
 };
 
 /**
@@ -410,5 +414,33 @@ function parseAroundLatLngFromString(value) {
   return {
     lat: parseFloat(pattern[1]),
     lng: parseFloat(pattern[2]),
+  };
+}
+
+function enhanceConfiguration(searchParametersFromUrl) {
+  return (configuration, widgetDefinition) => {
+    if (!widgetDefinition.getConfiguration) return configuration;
+
+    // Get the relevant partial configuration asked by the widget
+    const partialConfiguration = widgetDefinition.getConfiguration(
+      configuration,
+      searchParametersFromUrl
+    );
+
+    const customizer = (a, b) => {
+      // always create a unified array for facets refinements
+      if (Array.isArray(a)) {
+        return union(a, b);
+      }
+
+      // avoid mutating objects
+      if (isPlainObject(a)) {
+        return mergeWith({}, a, b, customizer);
+      }
+
+      return undefined;
+    };
+
+    return mergeWith({}, configuration, partialConfiguration, customizer);
   };
 }
